refactor: 소지금 단위 Integer -> double 수정, 주문 처리 시 소지금 차감

+++ src/main/java/com/sparta/gimbap_heaven/user/dto/UserResponseDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 public class UserResponseDto {
     private String username;
     private String email;
-    private Integer money;
+    private double money;
     private String intro;
 
     public UserResponseDto(User user){
